refactor: modulize ASF parser
